Move SearchableItem attribute formatting to separate presenter
