+++ src/schema/SearchableItem/index.ts
@@ -8,27 +8,7 @@ import { toGlobalId } from "graphql-relay"
 import { Searchable } from "schema/searchable"
 import { NodeInterface, GravityIDFields } from "schema/object_identification"
 import { ResolverContext } from "types/graphql"
-import { stripTags } from "lib/helpers"
-import moment from "moment"
-import { exhibitionPeriod } from "lib/date"
-
-const hrefFromAutosuggestResult = item => {
-  if (item.href) return item.href
-  switch (item.label) {
-    case "Profile":
-      return `/${item.id}`
-    case "Fair":
-      return `/${item.profile_id}`
-    case "Sale":
-      return `/auction/${item.id}`
-    case "City":
-      return `/shows/${item.id}`
-    case "MarketingCollection":
-      return `/collection/${item.id}`
-    default:
-      return `/${item.model}/${item.id}`
-  }
-}
+import { SearchableItemPresenter } from "./SearchableItemPresenter"
 
 export const SearchableItem = new GraphQLObjectType<any, ResolverContext>({
   name: "SearchableItem",
@@ -41,48 +21,7 @@ export const SearchableItem = new GraphQLObjectType<any, ResolverContext>({
     },
     description: {
       type: GraphQLString,
-      resolve: ({
-        description,
-        display,
-        end_at,
-        label,
-        published_at,
-        start_at,
-      }) => {
-        switch (label) {
-          case "Article":
-            let formattedPublishedAt
-            if (published_at) {
-              formattedPublishedAt = moment
-                .utc(published_at)
-                .format("MMM Do, YYYY")
-            }
-
-            if (published_at && description) {
-              return `${formattedPublishedAt} ... ${description}`
-            } else if (published_at) {
-              return formattedPublishedAt
-            } else {
-              return description
-            }
-          case "City":
-            return `Browse current exhibitions in ${display}`
-          case "Fair":
-          case "Sale":
-            const period = exhibitionPeriod(start_at, end_at)
-
-            return `Sale running from ${period}`
-          case "Show":
-          case "Booth":
-            return ""
-          default:
-            if (!description || description.length === 0) {
-              return null
-            } else {
-              return stripTags(description)
-            }
-        }
-      },
+      resolve: item => new SearchableItemPresenter(item).formattedDescription(),
     },
     displayLabel: {
       type: GraphQLString,
@@ -94,40 +33,11 @@ export const SearchableItem = new GraphQLObjectType<any, ResolverContext>({
     },
     href: {
       type: GraphQLString,
-      resolve: item => hrefFromAutosuggestResult(item),
+      resolve: item => new SearchableItemPresenter(item).href(),
     },
     searchableType: {
       type: GraphQLString,
-      resolve: ({ label, owner_type }) => {
-        switch (label) {
-          case "Profile":
-            const institutionTypes = [
-              "PartnerInstitution",
-              "PartnerInstitutionalSeller",
-            ]
-            if (institutionTypes.includes(owner_type)) {
-              return "Institution"
-            } else if (owner_type === "FairOrganizer") {
-              return "Fair"
-            } else {
-              return "Gallery"
-            }
-          case "Gene":
-            return "Category"
-
-          // TODO: How do we correctly display Sale/Auction types?
-          // There's nothing to distinguish the two types present
-          // in the special `match` JSON returned from the Gravity API.
-          case "Sale":
-            return "Auction"
-
-          case "MarketingCollection":
-            return "Collection"
-
-          default:
-            return label
-        }
-      },
+      resolve: item => new SearchableItemPresenter(item).searchableType(),
     },
   },
 })
